# optimize webshop building time and bundle size