--- a/react/assets/forus-webshop/scss/_common/includes/fonts.scss
+++ b/react/assets/forus-webshop/scss/_common/includes/fonts.scss
@@ -1,3 +1,0 @@
-@each $font in $fonts {
-    @import url($font);
-}