--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-2fa-restriction.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-2fa-restriction.scss
@@ -1,4 +1,4 @@
-&.block-2fa-restriction {
+.block.block-2fa-restriction {
     display: flex;
     flex-direction: column;
     margin: 0 0 20px;