--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-about-me_app.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-about-me_app.scss
@@ -1,4 +1,4 @@
-&.block-about-me_app {
+.block.block-about-me_app {
     display: flex;
     flex-direction: row;
     padding-top: 60px;