--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-accessibility.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-accessibility.scss
@@ -1,4 +1,4 @@
-&.block-accessibility {
+.block.block-accessibility {
     width: 100%;
     padding: 32px;
     background: #fff;