--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-action-card.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-action-card.scss
@@ -1,4 +1,4 @@
-&.block-action-card {
+.block.block-action-card {
     display: flex;
     background: #ffffff;
     padding: 25px;