--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-announcement.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-announcement.scss
@@ -1,4 +1,4 @@
-&.block-announcement {
+.block.block-announcement {
     width: 100%;
     padding: 15px 25px;
     background: #ffffff;