--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-app_download.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-app_download.scss
@@ -1,4 +1,4 @@
-&.block-app_download {
+.block.block-app_download {
     .app_download-row {
         display: flex;
         flex-direction: row;