--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-app_links.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-app_links.scss
@@ -1,4 +1,4 @@
-&.block-app_links {
+.block.block-app_links {
     display: flex;
     margin: 0 0 20px;
     flex-direction: row;