--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-auth-2fa.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-auth-2fa.scss
@@ -1,4 +1,4 @@
-&.block-auth-2fa {
+.block.block-auth-2fa {
     --success-icon-color: #39b924;
     margin-bottom: 20px;
     background-color: #fff;