--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-bookmark-toggle.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-bookmark-toggle.scss
@@ -1,4 +1,4 @@
-&.block-bookmark-toggle {
+.block.block-bookmark-toggle {
     width: 40px;
     height: 40px;
     border-radius: 20px;