--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-card-table.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-card-table.scss
@@ -1,4 +1,4 @@
-&.block-card-table {
+.block.block-card-table {
     table {
         border-collapse: collapse;
         width: 100%;