--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-cms-funds.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-cms-funds.scss
@@ -1,4 +1,4 @@
-&.block-cms-funds {
+.block.block-cms-funds {
     width: 100%;
 
     .fund-item {