--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-cms.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-cms.scss
@@ -1,3 +1,3 @@
-&.block-cms{
+.block.block-cms{
     display: block;
 }