--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-data-list.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-data-list.scss
@@ -1,4 +1,4 @@
-&.block-data-list {
+.block.block-data-list {
     display: table;
     border-collapse: collapse;