--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-empty.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-empty.scss
@@ -1,4 +1,4 @@
-&.block-empty {
+.block.block-empty {
     background: #fff;
     border-radius: var(--border-radius);
     box-shadow: 0 5px 30px rgba(0, 0, 0, .05);