--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-error-page.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-error-page.scss
@@ -1,4 +1,4 @@
-&.block-error-page {
+.block.block-error-page {
     display: flex;
     flex-direction: column;
     height: 100vh;