--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-pre-check-result.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-pre-check-result.scss
@@ -1,4 +1,4 @@
-&.block-fund-pre-check-result {
+.block.block-fund-pre-check-result {
     .fund-item {
         display: grid;
         overflow: hidden;