--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-pre-check.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-pre-check.scss
@@ -1,4 +1,4 @@
-&.block-fund-pre-check {
+.block.block-fund-pre-check {
     .pre-check-info {
         .pre-check-info-title {
             font: 700 22px/32px var(--base-font);