--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-request.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-request.scss
@@ -1,4 +1,4 @@
-&.block-fund-request {
+.block.block-fund-request {
     .card {
         border: 1px solid var(--border-color);