--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-requests.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-fund-requests.scss
@@ -1,4 +1,4 @@
-&.block-fund-requests {
+.block.block-fund-requests {
     margin: 0 0 20px;
 
     .fund-request-item {