--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-funds-list.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-funds-list.scss
@@ -1,4 +1,4 @@
-&.block-funds-list {
+.block.block-funds-list {
     .fund-item {
         display: grid;
         position: relative;