--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-info-box.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-info-box.scss
@@ -1,4 +1,4 @@
-&.block-info-box {
+.block.block-info-box {
     background-color: #FFFDEB;
     border: 1px solid var(--border-color);
     padding: 15px;