--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-key-value-list.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-key-value-list.scss
@@ -1,4 +1,4 @@
-&.block-key-value-list {
+.block.block-key-value-list {
     display: flex;
     flex-direction: column;
     gap: 10px;