--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-label-tabs.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-label-tabs.scss
@@ -1,4 +1,4 @@
-&.block-label-tabs {
+.block.block-label-tabs {
     display: flex;
     margin-right: 20px;
     border-top-left-radius: var(--label-set-border-radius);