--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-loader.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-loader.scss
@@ -1,4 +1,4 @@
-&.block-loader {
+.block.block-loader {
     display: flex;
     margin: 0 0 20px;
     flex-direction: column;