--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-map-office-card.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-map-office-card.scss
@@ -1,4 +1,4 @@
-&.block-map-office-card {
+.block.block-map-office-card {
     border-radius: var(--border-radius);
     box-shadow: 2px 2px 15px rgba($color: #000000, $alpha: 0.05);
     background: #fff;