--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-mobile-menu.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-mobile-menu.scss
@@ -1,6 +1,6 @@
 $footer_base_font: 'Open Sans', arial, sans-serif;
 
-&.block-mobile-menu {
+.block.block-mobile-menu {
     --item-height: 44px;
     --item-line-height: 18px;
     position: fixed;