--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-navbar-search.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-navbar-search.scss
@@ -1,4 +1,4 @@
-&.block-navbar-search {
+.block.block-navbar-search {
     display: flex;
     flex-grow: 1;
     position: relative;