--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-notifications-table.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-notifications-table.scss
@@ -1,4 +1,4 @@
-&.block-notifications-table {
+.block.block-notifications-table {
     .notification-date {
         flex-direction: row;
         font: 400 13px/26px var(--base-font);