--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-office.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-office.scss
@@ -1,4 +1,4 @@
-&.block-office {
+.block.block-office {
     width: 100%;
     margin: 0 0 30px;