--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-offices.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-offices.scss
@@ -1,4 +1,4 @@
-&.block-offices {
+.block.block-offices {
     border: var(--showcase-item-border);
     box-shadow: var(--showcase-item-shadow);
     background: var(--showcase-item-background);