--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-organization-info.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-organization-info.scss
@@ -1,4 +1,4 @@
-&.block-organization-info {
+.block.block-organization-info {
     display: block;
     width: 100%;
     padding: 20px 0 20px 0;