--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-organizations.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-organizations.scss
@@ -1,4 +1,4 @@
-&.block-organizations {
+.block.block-organizations {
     margin-bottom: 30px;
 
     .organization-item {