--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-pane.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-pane.scss
@@ -1,4 +1,4 @@
-&.block-pane {
+.block.block-pane {
     box-shadow: var(--box-shadow);
     margin-bottom: 30px;
     border-radius: var(--border-radius);