--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-pincode.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-pincode.scss
@@ -1,4 +1,4 @@
-&.block-pincode {
+.block.block-pincode {
     margin: 0 auto 20px;
     padding: 0;
     text-align: center;