--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-pre-check-banner.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-pre-check-banner.scss
@@ -1,4 +1,4 @@
-&.block-pre-check-banner {
+.block.block-pre-check-banner {
     display: flex;
     border: 1px solid var(--border-color);
     border-radius: var(--border-radius);