--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-products-list.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-products-list.scss
@@ -1,4 +1,4 @@
-&.block-products-list {
+.block.block-products-list {
     --products-title-color: var(--tc);
     --products-text-color: var(--tc);
     --products-gap: 15px;