--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-products.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-products.scss
@@ -1,4 +1,4 @@
-&.block-products {
+.block.block-products {
     --products-columns: 3;
     --products-gap: 20px;
     --products-content-padding: 15px;