--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-provider.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-provider.scss
@@ -1,4 +1,4 @@
-&.block-provider {
+.block.block-provider {
     width: 100%;
     margin: 0 0 30px;
     background-color: #fff;