--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-reimbursement.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-reimbursement.scss
@@ -1,4 +1,4 @@
-&.block-reimbursement {
+.block.block-reimbursement {
     display: flex;
     flex-direction: column;
     cursor: default;