--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-reimbursements.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-reimbursements.scss
@@ -1,4 +1,4 @@
-&.block-reimbursements {
+.block.block-reimbursements {
     margin-bottom: 30px;
 
     .reimbursement-item {