--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-reservation.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-reservation.scss
@@ -1,4 +1,4 @@
-&.block-reservation {
+.block.block-reservation {
     display: flex;
     flex-direction: column;
     cursor: default;