--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-reservations.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-reservations.scss
@@ -1,4 +1,4 @@
-&.block-reservations {
+.block.block-reservations {
     display: flex;
     flex-direction: column;
     gap: 15px;