--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-schedule-list.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-schedule-list.scss
@@ -1,4 +1,4 @@
-&.block-schedule-list {
+.block.block-schedule-list {
     display: flex;
     flex-direction: column;
     flex-grow: 1;