--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-search-results.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-search-results.scss
@@ -1,4 +1,4 @@
-&.block-search-results {
+.block.block-search-results {
     .search-wrapper {
         display: flex;
         width: 100%;