--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-sessions.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-sessions.scss
@@ -1,4 +1,4 @@
-&.block-sessions {
+.block.block-sessions {
     display: flex;
     flex-direction: column;