--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-showcase.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-showcase.scss
@@ -1,4 +1,4 @@
-&.block-showcase {
+.block.block-showcase {
     --aside-width: 330px;
     position: relative;
     cursor: default;