--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-sign_up-provider.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-sign_up-provider.scss
@@ -1,4 +1,4 @@
-&.block-sign_up-provider {
+.block.block-sign_up-provider {
     display: flex;
     flex-direction: row;
     padding-top: 60px;