--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-sign_up.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-sign_up.scss
@@ -1,4 +1,4 @@
-&.block-sign_up {
+.block.block-sign_up {
     padding: 50px 0;
     cursor: default;