--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-steps.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-steps.scss
@@ -1,4 +1,4 @@
-&.block-steps {
+.block.block-steps {
     .block-steps-description {
         font: 400 16px/28px var(--base-font);
         text-align: center;