--- a/react/assets/forus-webshop/scss/_common/sections/blocks/block-voucher-records.scss
+++ b/react/assets/forus-webshop/scss/_common/sections/blocks/block-voucher-records.scss
@@ -1,4 +1,4 @@
-&.block-voucher-records {
+.block.block-voucher-records {
     display: flex;
     flex-direction: column;
     flex-grow: 1;