--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-action-card.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-action-card.scss
@@ -1,4 +1,4 @@
-&.block-action-card {
+.block.block--action-card {
     .block-card-details {
         .block-card-title {
             color: #000;