--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-breadcrumbs.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-breadcrumbs.scss
@@ -1,4 +1,4 @@
-&.block-breadcrumbs .breadcrumb-item {
+.block.block--breadcrumbs .breadcrumb-item {
     font-size: 14px;
     color: var(--color-primary);
     font-weight: 500;