--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-cms-funds.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-cms-funds.scss
@@ -1,4 +1,4 @@
-&.block-cms-funds {
+.block.block--cms-funds {
     width: 100%;
     padding-top: 50px;