--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-cms.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-cms.scss
@@ -1,4 +1,4 @@
-&.block-cms {
+.block.block--cms {
     padding: 70px 0;
 
     .fund-item {