--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-copyrights.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-copyrights.scss
@@ -1,4 +1,4 @@
-&.block-copyrights {
+.block.block--copyrights {
     background-color: #292F2E;
     box-shadow: inset 0px 1px 0px rgba(255, 255, 255, 0.15);
     border-top: none;