--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-empty.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-empty.scss
@@ -1,4 +1,4 @@
-&.block-empty {
+.block.block--empty {
     .block-description {
         font-weight: 500;
     }