--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-footer.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-footer.scss
@@ -1,4 +1,4 @@
-&.block-footer {
+.block.block--footer {
     position: relative;
     margin-bottom: 0;