--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-map.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-map.scss
@@ -1,4 +1,4 @@
-&.block-map {
+.block.block--map {
     .block-content {
         .block-title {
             font: 500 50px/68px var(--base-font);