--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-markdown.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-markdown.scss
@@ -1,4 +1,4 @@
-&.block-markdown {
+.block.block--markdown {
     h1, h2, h3, h4, h5, h6 {
         color: #003B49;
     }