--- a/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-navbar.scss
+++ b/react/assets/forus-webshop/scss/doetegoed/sections/blocks/block-navbar.scss
@@ -1,4 +1,4 @@
-&.block-navbar {
+.block.block-navbar {
     .navbar-list {
         .navbar-item-wrapper {
             .navbar-item {