## Changes description & testing suggestions
This update should reduce the building time and bundle size for webshops.

Locally I got this results:

### Before
```
[only general webshop]
    58sec => 30.8Mb [app.js file per webshop] or [4.2Mb as .gz]

[all webshops]
    205sec => 30.8Mb [app.js file per webshop] or [4.2Mb as .gz]
```

### After
```
[only general webshop]
    37sec => 5.9Mb [app.js file per webshop] or [1.2Mb as .gz]

[all webshops]
    145sec => 5.9Mb [app.js file per webshop] or [1.2Mb as .gz]
```

### Testing scope:
We need to test that the implementations look the same as before the update. For the testing I would open and compare side by side 5 pages: home, products, funds, fund-request and vouchers pages from each implementation.

Also please check that the fonts are the same descriptions and titles.

<!---
Add tag if PR:
- [ ] *Needs Translations* @lexlog @irinaBerendeeva87
- [ ] *Could affect implementations custom css* @lexlog @irinaBerendeeva87
-->

## Developers checklist
- [ ] *Check that dusk tests are working locally on compatible branch*
- [ ] *Mobile version of changes is developed* - if its a webshop feature, mobile version for this feature is developed

## QA checklist
- *Check regress in implementations custom css* - changes are not breaking other implementations designes
- Feature is tested in different screen sizes - desktop, mobile
- WCAG requirements are met - new feature is accessible by keyboard, there are an alt texts
- Translations are done